requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
alembic
psycopg2-binary
pydantic
python-dotenv
openai
slowapi
itsdangerous
